Set PROXY protocol for Service if set in Deployment
Check the current Deployment to determine whether PROXY protocol is enabled
on the router, and configure the load-balancer Service accordingly.

This change ensures that the Service and Deployment will be configured
harmoniously in the future when we start applying the patches from
UnsupportedExtensions to the Deployment.

* pkg/manifests/manifests.go (RouterServiceCloud): Add useProxyProtocol
parameter.
* pkg/manifests/manifests_test.go: Fix call to RouterServiceCloud.
* pkg/operator/controller/controller.go: Check the current Deployment's
environment variables to determine whether ROUTER_USE_PROXY_PROTOCOL is
set, and call RouterServiceCloud accordingly.