Avoid dropping traffic during upgrade
Omit the affinity policy when using host network.  The scheduler already
prevents colocation of pods that use host networking and request the same
ports, so the affinity policy is superfluous in this case.

Change the affinity policy and deployment strategy when using the
load-balancer or private endpoint publishing strategy as follows.  First,
configure affinity for replicas of different generations of same
controller.  Second, configure anti-affinity for replicas of same
generation of same controller.  Finally, configure the deployment strategy
to surge.

The intention of the changes for the load-balancer and private endpoint
publishing strategies is to avoid dropping traffic during rolling updates
of an ingress controller's deployment.  If a node loses local endpoints for
the deployment, then the service proxy will drop traffic to that node for
that ingress controller, and it may take some time for the load balancer to
stop sending traffic to that node.  These changes, combined with a change
to the ReplicaSet controller, will ensure that a node that has local
endpoints for an ingress controller at the start of a rollout will continue
to have local endpoints during and at completion of the rollout, thus
preventing traffic from being dropped.

This commit is related to bug 1709958.

https://bugzilla.redhat.com/show_bug.cgi?id=1709958

* pkg/operator/controller/controller_router_deployment.go
(desiredRouterDeployment): Set the deployment's pod selector to a copy of
the pod template's labels, so that subsequently mutating labels does not
mutate the selector.  If the ingress controller uses the host network, do
not set any affinity policy.  If the ingress controller uses the
load-balancer or private endpoint publishing strategy, add a "hash" label
to identify the deployment's generation, configure affinity for replicas of
different generations of the same ingress controller, configure
anti-affinity for the same generation of the same ingress controller, and
configure the deployment strategy to surge.
(deploymentHash): New function.  Return a stringified hash value for the
given deployment, using only the fields that, if changed, should trigger an
update.
(hashableDeployment): New function.  Return a copy of the given deployment
with fields that should be used for computing its hash copied over, fields
that are slices sorted, and fields that should be ignored zeroed.
(deepHashObject): New function.  Hash the given object.
(deploymentConfigChanged): Compare hashes of the current and expected
deployments instead of comparing fields directly.  Set labels during an
update.
(cmpEnvs, cmpVolumes, cmpSecretVolumeSource, cmpTolerations): Deleted.
* pkg/operator/controller/controller_router_deployment_test.go
(TestDesiredRouterDeployment): Verify that the hash label is set when it
should be and that it has the right hash value.
(TestDeploymentConfigChanged): Add test cases to verify that the hash is
ignored in the deployment labels and affinity.  Add a test case to verify
that deleting the affinity policy is not ignored.  Add test cases to verify
that ordering of the label selector expressions in affinity terms is
ignored.
* pkg/operator/controller/names.go (ControllerDeploymentHashLabel):
New constant.