hashableDeployment: Fix liveness/readiness probes
When comparing deployments to determine whether an update is required,
treat implicit and explicit default values as equal.

Before this commit, the update logic would keep trying to revert the
default values that the API set for probe parameters.

* pkg/operator/controller/ingress/deployment.go (hashableDeployment): Use
the new hashableProbe function so that two deployments have the same hash
if the only difference between them is that one does not specify liveness
or readiness probe parameters and the other specifies the default parameter
values.
(hashableProbe): New function.  Return a copy of the given probe, copying
exactly the fields that should be used for computing a deployment's hash,
and zeroing out any fields where the values are equal to their defaults.