ensureLoadBalancerService: Update service if needed
If a load-balancer service already exists but is not the same as the
desired load-balancer service, update the current service.

* pkg/operator/controller/ingress/load_balancer_service.go
(ensureLoadBalancerService): If a service exists and none is desired,
delete the existing service.  If a service is desired and one exists but
does not match the desired service, update it using
updateLoadBalancerService.
(updateLoadBalancerService, loadBalancerServiceChanged(): New methods, used
by ensureLoadBalancerService.