status: Watch clusteroperators
Add a watch on clusteroperators in the status controller so that it updates
or recreates the ingress clusteroperator as necessary should something else
update or delete it.

* manifests/00-cluster-role.yaml: Allow the operator to list and watch
clusteroperators.
* pkg/operator/controller/status/controller.go (New): Watch
clusteroperators with a map function and predicate to reconcile
ingresscontrollers.operator.openshift.io/default if
clusteroperators.config.openshift.io/ingress is updated.
* test/e2e/operator_test.go (TestOperatorRecreatesItsClusterOperator): New
test.  Verify that the operator recreates the "ingress" clusteroperator if
the clusteroperator is deleted.