Allow TLSv1.3 and the "Modern" TLS profile
OpenShift 4.6 is built using UBI 8, which has OpenSSL 1.1.1, which supports
TLSv1.3, so allow specifying spec.tlsSecurityProfile.type: Modern or
spec.tlsSecurityProfile.custom.minTLSVersion: VersionTLS13.

* pkg/operator/controller/ingress/controller.go
tlsProfileSpecForSecurityProfile): Allow use of the "Modern" profile.
(validTLSVersions): Add configv1.VersionTLS13.
(validateTLSSecurityProfile): Verify that some non-TLSv1.3 cipher is
specified if minTLSVersion is a version below TLSv1.3, and verify that some
TLSv1.3 cipher is specified if minTLSVersion is TLSv1.3.
* pkg/operator/controller/ingress/controller_test.go
(TestTLSProfileSpecForSecurityProfile)
(TestTLSProfileSpecForIngressController): Verify that the "Modern" profile
is used if it is specified.
* pkg/operator/controller/ingress/deployment.go (desiredRouterDeployment)
(inferTLSProfileSpecFromDeployment): Allow TLSv1.3.
* pkg/operator/controller/ingress/deployment_test.go
(TestDesiredRouterDeployment): Expect SSL_MIN_VERSION to be set to TLSv1.3
if TLSv1.3 was specified on the ingresscontroller.
(TestInferTLSProfileSpecFromDeployment): Add test cases for
SSL_MIN_VERSION=TLSv1.2 and for SSL_MIN_VERSION=TLSv1.3.